refactor(frontline): extract shared bulk removal hook
